chore: fix refs

+++ collab-database/src/macros.rs
@@ -1,5 +1,3 @@
-use collab::preclude::Any;
-
 #[macro_export]
 macro_rules! impl_str_update {
   ($setter1: ident, $setter2: ident, $key: expr) => {
@@ -66,13 +64,17 @@ macro_rules! impl_i64_update {
 macro_rules! impl_i32_update {
   ($setter1: ident, $setter2: ident, $key: expr) => {
     pub fn $setter1(self, value: i32) -> Self {
-      self.map_ref.insert(self.txn, $key, Any::BigInt(value));
+      self
+        .map_ref
+        .insert(self.txn, $key, Any::BigInt(value as i64));
       self
     }
 
     pub fn $setter2(self, value: Option<i32>) -> Self {
       if let Some(value) = value {
-        self.map_ref.insert(self.txn, $key, Any::BigInt(value));
+        self
+          .map_ref
+          .insert(self.txn, $key, Any::BigInt(value as i64));
       }
       self
     }
@@ -83,13 +85,17 @@ macro_rules! impl_i32_update {
 macro_rules! impl_u8_update {
   ($setter1: ident, $setter2: ident, $key: expr) => {
     pub fn $setter1(self, value: u8) -> Self {
-      self.map_ref.insert(self.txn, $key, Any::BigInt(value));
+      self
+        .map_ref
+        .insert(self.txn, $key, Any::BigInt(value as i64));
       self
     }
 
     pub fn $setter2(self, value: Option<u8>) -> Self {
       if let Some(value) = value {
-        self.map_ref.insert(self.txn, $key, Any::BigInt(value));
+        self
+          .map_ref
+          .insert(self.txn, $key, Any::BigInt(value as i64));
       }
       self
     }
